he cambiado cosas porque tenia archivos que no estaban bien guardados,,ademas he ñpuesto para que funcione el ateque de disparo,el resto funcina pero no se ven en visual studio
