+++ archon.vcxproj.filters
@@ -38,12 +38,10 @@
     <ClInclude Include="include\juego.h" />
     <ClInclude Include="include\renderizador.h" />
     <ClInclude Include="include\tablero.h" />
-    <ClInclude Include="include\cabra.h" />
     <ClInclude Include="include\transicion.h" />
     <ClInclude Include="include\cursor.h" />
     <ClInclude Include="include\tarjetaID.h" />
     <ClInclude Include="include\cerdo.h" />
-    <ClInclude Include="include\gallina.h" />
     <ClInclude Include="include\granjero.h" />
     <ClInclude Include="include\llama.h" />
     <ClInclude Include="include\oveja.h" />
@@ -51,7 +49,13 @@
     <ClInclude Include="include\controles.h" />
     <ClInclude Include="include\jugador.h" />
     <ClInclude Include="include\estructuras.h" />
-    <ClInclude Include="include\interaccion.h" />
+    <ClInclude Include="include\ondaa.h" />
+    <ClInclude Include="include\ataquee.h" />
+    <ClInclude Include="include\cabraa.h" />
+    <ClInclude Include="include\disparoo.h" />
+    <ClInclude Include="include\golpee.h" />
+    <ClInclude Include="include\gallinaa.h" />
+    <ClInclude Include="include\interaccionn.h" />
   </ItemGroup>
   <ItemGroup>
     <None Include="docs\cuadernoBitacora.md">
@@ -79,7 +83,6 @@
     </ClCompile>
     <ClCompile Include="src\menu.cpp" />
     <ClCompile Include="src\animal.cpp" />
-    <ClCompile Include="src\arena.cpp" />
     <ClCompile Include="src\juego.cpp" />
     <ClCompile Include="src\renderizador.cpp" />
     <ClCompile Include="src\tablero.cpp" />
@@ -94,7 +97,9 @@
     <ClCompile Include="src\oveja.cpp" />
     <ClCompile Include="src\creditos.cpp" />
     <ClCompile Include="src\controles.cpp" />
-    <ClCompile Include="src\jugador.cpp" />
+    <ClCompile Include="src\ataque.cpp" />
+    <ClCompile Include="src\arenaa.cpp" />
+    <ClCompile Include="src\jugadorr.cpp" />
   </ItemGroup>
   <ItemGroup>
     <Library Include="lib\ETSIDI.lib">
