+++ include/cabraa.h
@@ -1,6 +1,6 @@
 #pragma once
 #include "animal.h"
-#include "onda.h"
+#include "ondaa.h"
 
 class Cabra : public Animal {
 
@@ -11,7 +11,7 @@ public:
         max_casillas_movidas_ = 4;
         nFrames = 5;
 
-        ataque_ = new Onda(7, 100.0f, 0.6f,"../assets/Sprites/creditos/botella.png", 48.0f, 0.4f,1.0f, 0.0f, 1.0f);
+        ataque_ = new Onda(7, 100.0f, 0.6f, "../assets/Sprites/creditos/botella.png", 48.0f, 0.4f, 1.0f, 0.0f, 1.0f);
     }
 
     const char* getTipoAtaque() const override { return "Onda"; }
