+++ include/gallinaa.h
@@ -1,7 +1,7 @@
 #pragma once
 #include "animal.h"
 #include "estructuras.h"
-#include "disparo.h"
+#include "disparoo.h"
 
 class Gallina : public Animal {
 
@@ -12,7 +12,7 @@ public:
         max_casillas_movidas_ = 2;
         nFrames = 8;
 
-        ataque_ = new Disparo(12,140.0f,0.8f,"../assets/Sprites/ataques/botella.png",20.0f,0.35f,1.0f, 1.0f, 0.0f);
+        ataque_ = new Disparo(12, 140.0f, 0.8f, "../assets/Sprites/creditos/botella.png", 20.0f, 0.35f, 1.0f, 1.0f, 0.0f);
     }
 
     const char* getTipoAtaque() const override { return "Golpe"; }
