+++ include/juego.h
@@ -5,9 +5,9 @@
 #include "arena.h"
 #include "renderizador.h" 
 #include "animal.h"
-#include "cabra.h"
+#include "cabraa.h"
 #include "cerdo.h"
-#include "gallina.h"
+#include "gallinaa.h"
 #include "oveja.h"
 #include "granjero.h"
 #include "transicion.h"
